- Migrated custom monosaccharide form to be a tab of the main window form - Deleted custom monosaccharide form
